fix: 修复心率统计类型推断错误-连续3次CI编译失败
根因：minRR/maxRR 用 "Int? ?: Double" 兜底，Kotlin 推断为 Number 公共父类型
（Int 和 Double 的 LUB），导致 60000.0 / minRR 找不到 div(Number) 重载。

错误：
e: EcgFeatureExtractor.kt: None of the following candidates is applicable:
  div(Byte/Double/Float/Int/Long/Short)

修复：统一转 Double
  val minRR = (rrForRange.minOrNull() ?: medianRR.toInt()).toDouble()
  val maxRR = (rrForRange.maxOrNull() ?: medianRR.toInt()).toDouble()

此 bug 从 5efd6b6 引入，导致连续 3 次 CI 编译失败（5efd6b6/6769660/ef7b9aa）。
算法逻辑不变，仅修复类型。

+++ shared/src/main/java/com/wearhealth/companion/shared/EcgFeatureExtractor.kt
@@ -381,8 +381,8 @@ object EcgFeatureExtractor {
         val normalRRs = rrIntervals.filter { it.toDouble() in lowerBound..upperBound }
         val rrForRange = if (normalRRs.size >= 3) normalRRs else rrIntervals
 
-        val minRR = rrForRange.minOrNull() ?: medianRR
-        val maxRR = rrForRange.maxOrNull() ?: medianRR
+        val minRR = (rrForRange.minOrNull() ?: medianRR.toInt()).toDouble()
+        val maxRR = (rrForRange.maxOrNull() ?: medianRR.toInt()).toDouble()
         return HeartRateStats(
             avgHr = medianHR,  // 中位数反算，抗误检
             minHr = (60000.0 / maxRR).toInt(),  // 长 RR = 低心率
